requirements.txt
# RAG System Dependencies
requests>=2.31.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
streamlit>=1.31.0
python-dotenv>=1.0.0
